move version control to features
