+++ src/components/table/UpdatesDialog.tsx
@@ -4,7 +4,7 @@ import {
     generateChangeRecords,
     generateFileLinkKey,
     getTrackedFileLink,
-} from "@src/content/versionControl"
+} from "@src/features/versionControl"
 import { FileLink } from "@src/types/pageContentTypes"
 
 interface UpdatesDialogProps {
