+++ src/content/parsers/table/parseFileLinkRows.ts
@@ -3,7 +3,7 @@ import parseFileSize from "@src/content/parsers/parseFileSize"
 import { FileLink, Image } from "@src/types/pageContentTypes"
 import splitFileName from "@src/content/parsers/splitFileName"
 import { getTimeDifferenceString } from "@src/content/enhancers/getTimeDifferenceString"
-import { extractUrlSegments } from "@src/content/versionControl"
+import { extractUrlSegments } from "@src/features/versionControl"
 
 const IMAGE_FILE_EXTENSIONS = ["png", "jpg", "jpeg", "gif", "webp", "svg"]
