+++ src/pages/content/pages/NotFound.tsx
@@ -8,7 +8,7 @@ import QuickLinkCard from "@src/components/quickAccess/QuickLinkCard"
 import QuickLinkContainer from "@src/components/quickAccess/QuickLinkContainer"
 import { getQuickLinks } from "@src/components/quickAccess/QuickLinkManager"
 import Title from "@src/components/typography/Title"
-import { BASE_URL } from "@src/content/versionControl"
+import { BASE_URL } from "@src/features/versionControl"
 import { ConfigContext } from "@src/contexts/ConfigContext"
 import { QuickLink } from "@src/types/quickLinkTypes"
 import { useQuery } from "@tanstack/react-query"
